consolidate I/SearchService
